# Conflicts:
#	Style Guide.txt
#	src/main/java/Instances/InstanceBuilder_BGU.java
#	src/main/java/Instances/InstanceBuilder_MovingAI.java
#	src/main/java/Instances/InstanceManager.java
#	src/main/java/Instances/MAPF_Instance.java
#	src/main/java/Instances/Maps/Enum_MapCellType.java
#	src/main/java/Instances/Maps/GraphMap.java
#	src/main/java/Instances/Maps/GraphMapCell.java
#	src/main/java/Instances/Maps/I_Map.java
#	src/main/java/Instances/Maps/I_MapCell.java
#	src/main/java/Instances/Maps/MapFactory.java

+++ src/main/java/Instances/Maps/I_Coordinate.java
@@ -2,6 +2,21 @@ package Instances.Maps;
 
 public interface I_Coordinate {
 
+    /**
+     * Returns the euclidean distance to another {@link I_Coordinate coordinate}. Should return 0 iff this.equals(other)
+     * return true.  If other is null, or is not of the same runtime type as this, returns -1.
+     * @param other a {@link I_Coordinate coordinate}.
+     * @return the euclidean distance to another {@link I_Coordinate coordinate}. If other is null, or is not of the
+     * same runtime type as this, returns -1.
+     */
     float euclideanDistance(I_Coordinate other);
-    float manhattanDistance(I_Coordinate other);
+
+    /**
+     * Returns the manhattan distance to another {@link I_Coordinate coordinate}. Should return 0 iff this.equals(other)
+     * return true.  If other is null, or is not of the same runtime type as this, returns -1.
+     * @param other a {@link I_Coordinate coordinate}.
+     * @return the manhattan distance to another {@link I_Coordinate coordinate}.  If other is null, or is not of the
+     * same runtime type as this, returns -1.
+     */
+    int manhattanDistance(I_Coordinate other);
 }
